pyproject.toml
[project]
name = "titan"
version = "0.1.0"
description = "AI-Powered Options Trading System for Interactive Brokers"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.9.0",
    "anthropic>=0.40.0",
    "apscheduler>=3.10.0",
    "asyncpg>=0.29.0",
    "fastapi>=0.110.0",
    "hmmlearn>=0.3.0",
    "ib-async>=2.1.0",
    "langgraph>=0.2.0",
    "lightgbm>=4.3.0",
    "numpy>=1.26.0",
    "optuna>=4.7.0",
    "pandas>=2.2.0",
    "pandas-ta>=0.4.71b0",
    "prometheus-client>=0.20.0",
    "psycopg2-binary>=2.9.9",
    "py-vollib>=1.0.1",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "python-telegram-bot>=21.0",
    "pyyaml>=6.0.0",
    "quantstats>=0.0.62",
    "redis>=5.0.0",
    "river>=0.21.0",
    "scikit-learn>=1.4.0",
    "scipy>=1.12.0",
    "structlog>=24.1.0",
    "twilio>=9.0.0",
    "uvicorn>=0.27.0",
    "websockets>=12.0",
    "xgboost>=2.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.15.2",
]
